pyproject.toml
[tool.poetry]
name = "mcp-routing"
version = "0.1.0"
description = "MCP service for natural language routing using DeepSeek and OSM-based routing engines"
authors = ["Lu Liu <nudtlliu@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.110.0"
uvicorn = "^0.27.0"
pydantic = "^2.6.0"
requests = "^2.31.0"
folium = "^0.15.0"
python-dotenv = "^1.0.0"
polyline = "^2.0.0"
openai = "^1.12.0"
transformers = "^4.38.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
